More options (AddAuthorizationHeader, UseDateTimeOffset, MethodReturnType & AppendAsync)

+++ examples/RestEaseClientGeneratorConsoleApp/PetStore/Models/User.cs
@@ -1,10 +1,11 @@
+using System;
 using System.Collections.Generic;
 
 namespace RestEaseClientGeneratorConsoleApp.PetStore.Models
 {
     public class User
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Username { get; set; }
